Koios Alignment with 1.1.1 Bump Version to 1.18.2

+++ src/main/java/rest/koios/client/backend/api/pool/model/PoolInfo.java
@@ -112,7 +112,7 @@ public class PoolInfo {
     /**
      * Pool relative active stake share
      */
-    private Long sigma;
+    private Double sigma;
 
     /**
      * Total pool blocks on chain
